+++ benchmark.py
@@ -209,7 +209,7 @@ def parse_and_sort_unicode_piecemeal():
 
 @mark
 def full():
-    return TimeLog(Settings().get_timelog_file(), Settings().virtual_midnight).items
+    return TimeLog(Settings().get_timelog_file(), Settings().virtual_midnight, Settings().rounding_time).items
 
 
 def main():